# core-banking-psdx-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-banking-psdx
  description: PSD2/PSD3 and FIDA Regulatory Compliance Service
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/psdx
    description: Development Environment
  - url: /
    description: Local Development Environment
security:
  - apiKey: []
  - bearer: []
...
paths:
...
  /api/v1/sca/validate:
    post:
      tags:
        - Strong Customer Authentication
      summary: Validate SCA
      description: Validates Strong Customer Authentication for a customer
      operationId: validateSCA
      parameters:
        - name: X-Request-ID
          in: header
          description: Unique ID for the request
          schema:
            type: string
          example: 99391c7e-ad88-49ec-a2ad-99ddcb1f7721
        - name: X-Consent-ID
          in: header
          description: ID of the consent for the operation
          schema:
            type: string
          example: '12345'
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PSDSCAValidationRequestDTO'
        required: true
      responses:
        '200':
          description: SCA validated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDSCAValidationResponseDTO'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDSCAValidationResponseDTO'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDSCAValidationResponseDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDSCAValidationResponseDTO'
  /api/v1/sca/authenticate:
    post:
      tags:
        - Strong Customer Authentication
      summary: Initiate SCA
      description: Initiates Strong Customer Authentication for a customer
      operationId: initiateSCA
      parameters:
        - name: X-Request-ID
          in: header
          description: Unique ID for the request
          schema:
            type: string
          example: 99391c7e-ad88-49ec-a2ad-99ddcb1f7721
        - name: X-Consent-ID
          in: header
          description: ID of the consent for the operation
          schema:
            type: string
          example: '12345'
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PSDSCAAuthenticationRequestDTO'
        required: true
      responses:
        '200':
          description: SCA initiated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDSCAAuthenticationResponseDTO'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDSCAAuthenticationResponseDTO'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDSCAAuthenticationResponseDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDSCAAuthenticationResponseDTO'
...
  /api/auth/token:
    post:
      tags:
        - Authentication
      summary: Authenticate
      description: Authenticate a user and generate JWT tokens
      operationId: authenticate
      parameters:
        - name: X-Request-ID
          in: header
          description: Unique ID for the request
          schema:
            type: string
          example: 99391c7e-ad88-49ec-a2ad-99ddcb1f7721
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PSDAuthenticationRequestDTO'
        required: true
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDAuthenticationResponseDTO'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDAuthenticationResponseDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDAuthenticationResponseDTO'
...
  /api/v1/sca/required:
    get:
      tags:
        - Strong Customer Authentication
      summary: Check if SCA is required
      description: Checks if Strong Customer Authentication is required for a payment
      operationId: isSCARequired
      parameters:
        - name: amount
          in: query
          description: Payment amount
          required: true
          schema:
            type: number
            format: double
        - name: currency
          in: query
          description: Payment currency
          required: true
          schema:
            type: string
        - name: X-Request-ID
          in: header
          description: Unique ID for the request
          schema:
            type: string
          example: 99391c7e-ad88-49ec-a2ad-99ddcb1f7721
        - name: X-Consent-ID
          in: header
          description: ID of the consent for the operation
          schema:
            type: string
          example: '12345'
      responses:
        '200':
          description: Check completed successfully
          content:
            application/json:
              schema:
                type: boolean
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: boolean
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: boolean
...
components:
  schemas:
...
    PSDSCAValidationRequestDTO:
      required:
        - authenticationCode
        - challengeId
      type: object
      properties:
        challengeId:
          maxLength: 50
          minLength: 0
          pattern: ^[A-Za-z0-9\-_]+$
          type: string
          description: ID of the SCA challenge
          example: sca-123456
        authenticationCode:
          maxLength: 10
          minLength: 4
          pattern: ^[0-9]+$
          type: string
          description: Authentication code provided by the customer
          example: '123456'
      description: SCA validation request
    PSDSCAValidationResponseDTO:
      required:
        - success
      type: object
      properties:
        success:
          type: boolean
          description: Whether the validation was successful
          example: true
        authenticationToken:
          maxLength: 2048
          minLength: 10
          pattern: ^[A-Za-z0-9\-._~+/]+=*$
          type: string
          description: Authentication token for the validated session
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        expiresIn:
          type: integer
          description: Expiry time of the token in seconds
          format: int32
          example: 3600
        errorMessage:
          maxLength: 500
          minLength: 0
          type: string
          description: Error message if validation failed
          example: Invalid authentication code
      description: PSD SCA Validation response
    PSDSCAAuthenticationRequestDTO:
      required:
        - partyId
        - resourceId
        - resourceType
      type: object
      properties:
        partyId:
          type: string
          description: ID of the customer
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        resourceId:
          maxLength: 100
          minLength: 0
          pattern: ^[A-Za-z0-9\-_]+$
          type: string
          description: ID of the resource being accessed
          example: payment-123456
        resourceType:
          pattern: ^(PAYMENT|ACCOUNT|CONSENT|FUNDS_CONFIRMATION)$
          type: string
          description: Type of the resource being accessed
          example: PAYMENT
        amount:
          type: number
          description: Amount of the transaction
          format: double
          example: 100
        currency:
          type: string
          description: Currency of the transaction
          example: EUR
        preferredMethod:
          type: string
          description: Preferred authentication method
          example: SMS
      description: SCA authentication request
    PSDSCAAuthenticationResponseDTO:
      required:
        - challengeId
        - method
      type: object
      properties:
        challengeId:
          maxLength: 50
          minLength: 0
          pattern: ^[A-Za-z0-9\-_]+$
          type: string
          description: ID of the SCA challenge
          example: sca-123456
        method:
          pattern: ^(SMS|EMAIL|PUSH|BIOMETRIC|HARDWARE_TOKEN|SOFTWARE_TOKEN)$
          type: string
          description: Authentication method used
          example: SMS
        maskedTarget:
          maxLength: 100
          minLength: 0
          type: string
          description: Masked target of the authentication
          example: +49 *** *** 789
        expiresIn:
          type: integer
          description: Expiry time of the challenge in seconds
          format: int32
          example: 300
        additionalInfo:
          maxLength: 500
          minLength: 0
          type: string
          description: Additional information about the challenge
          example: An SMS has been sent to your registered mobile number
      description: PSD SCA Authentication response
...
    PSDAuthenticationRequestDTO:
      required:
        - password
        - username
      type: object
      properties:
        username:
          maxLength: 100
          minLength: 0
          type: string
          description: Username
          example: user@example.com
        password:
          maxLength: 128
          minLength: 8
          type: string
          description: Password
          example: password
      description: Authentication request
    PSDAuthenticationResponseDTO:
      required:
        - accessToken
        - tokenType
      type: object
      properties:
        accessToken:
          maxLength: 2048
          minLength: 10
          pattern: ^[A-Za-z0-9\-._~+/]+=*$
          type: string
          description: Access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refreshToken:
          maxLength: 2048
          minLength: 10
          pattern: ^[A-Za-z0-9\-._~+/]+=*$
          type: string
          description: Refresh token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        tokenType:
          pattern: ^(Bearer|Basic)$
          type: string
          description: Token type
          example: Bearer
        expiresIn:
          type: integer
          description: Expires in (seconds)
          format: int64
          example: 3600
      description: PSD Authentication response
...
  securitySchemes:
    apiKey:
      type: apiKey
      description: API key for TPP authentication
      name: X-API-KEY
      in: header
    bearer:
      type: http
      description: JWT token for user authentication
      scheme: bearer
      bearerFormat: JWT